refactor(dashboard): extract the session hooks out of the header
The header held the query, the query client, the router and the sign-out
mutation to render a name and a button. Reading the session off the router
context the guard already resolved also drops the second fetch of it.
